# openapi.yaml
components:
  schemas:
    EncryptedPayload:
      description: 'Generic model for an encrypted payload.


        Can correspond to current/legacy upload metadata or a file secret.'
      properties:
        payload:
          title: Payload
          type: string
      required:
      - payload
      title: EncryptedPayload
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: A service to ingest s3 file upload metadata produced by thedata-steward-kit
    upload command
  title: File Ingest Service
  version: 1.1.2
openapi: 3.1.0
paths:
  /federated/ingest_metadata:
    post:
      description: Decrypt payload, process metadata, file secret id and send success
        event
      operationId: ingestFileUploadMetadata
      parameters:
      - in: query
        name: service_config
        required: true
        schema:
          properties:
            config_yaml:
              anyOf:
              - format: path
                type: string
              - type: 'null'
              title: Config Yaml
          title: Service Config
          type: object
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncryptedPayload'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema: {}
          description: Received and decrypted data successfully.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Processes encrypted output data from the S3 upload script and ingests
        it into the Encryption Key Store, Internal File Registry and Download Controller.
      tags:
      - FileIngestService
  /federated/ingest_secret:
    post:
      description: Decrypt payload and deposit file secret in exchange for a secret
        id
      operationId: ingestSecret
      parameters:
      - in: query
        name: service_config
        required: true
        schema:
          properties:
            config_yaml:
              anyOf:
              - format: path
                type: string
              - type: 'null'
              title: Config Yaml
          title: Service Config
          type: object
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncryptedPayload'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Received and stored secret successfully.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Store file encryption/decryption secret and return secret ID.
      tags:
      - FileIngestService
  /health:
    get:
      description: Used to test if this service is alive
      operationId: health_health_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: health
      tags:
      - FileIngestService
  /legacy/ingest:
    post:
      deprecated: true
      description: Decrypt payload, process metadata, file secret and send success
        event
      operationId: ingestLegacyFileUploadMetadata
      parameters:
      - in: query
        name: service_config
        required: true
        schema:
          properties:
            config_yaml:
              anyOf:
              - format: path
                type: string
              - type: 'null'
              title: Config Yaml
          title: Service Config
          type: object
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncryptedPayload'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema: {}
          description: Received and decrypted data successfully.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Processes encrypted output data from the S3 upload script and ingests
        it into the Encryption Key Store, Internal File Registry and Download Controller.
      tags:
      - FileIngestService
tags:
- name: FileIngestService